backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "back-end for web-app"
package-mode = false
authors = [
    {name = "Ollldman",email = "dodrolla@gmail.com"}
]
requires-python = "^3.13"
dependencies = [
    "fastapi[standard] (>=0.135.0,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "alembic (>=1.18.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
